requirements.txt
mss
opencv-python
pytesseract
sounddevice
vosk
numpy
pillow
pyyaml
wavio